*** empty log message ***

+++ include/job.h
@@ -1,4 +1,4 @@
-//  $Id: job.h,v 1.59 2008/01/11 14:40:59 choutko Exp $
+//  $Id: job.h,v 1.60 2008/02/13 20:07:51 choutko Exp $
 // Author V. Choutko 24-may-1996
 //
 // Mar 24, 1997. ak. add _eventRtype 
@@ -165,7 +165,7 @@ const char *GetNtuplePath(){return _hextname;}
 void SetRootPath(const char * hname){_rextname=hname;}
 const char*  GetRootPath(){return _rextname;}
 static AMSNodeMap JobMap;
-static long GetNtupleFileSize();
+static long long GetNtupleFileSize();
 static time_t GetNtupleFileTime();
 
 #ifdef __DB__
